Fix: assert == '.git' error
If the '.git' is somewhere in the directories, even it is not configured
in gitosis config, when it walks through all directory names, it will
emit this assert error. The reason is that 'os.path.splitext(repo)' does
not work as expected, as it returns '.git' and '' for the input '.git'.